debugger: indicate server is ending
Currently, we say "listening" when we are ending the server. Change it
to "ending".

Fixes: https://github.com/nodejs/node/issues/39272

PR-URL: https://github.com/nodejs/node/pull/39334
